fix selecting multiple nested groupby keys
