+++ entities/api/groupby_helpers.py
@@ -1,21 +1,11 @@
 import typing
-from typing import TYPE_CHECKING, Annotated, Any, Optional, Sequence, Callable
+from typing import Annotated, Any, Optional
 import strawberry
-import database.models as db
 
-E = typing.TypeVar("E", db.File, db.Entity)
-T = typing.TypeVar("T")
 
-if TYPE_CHECKING:
-    from api.types.consensus_genome import ConsensusGenomeGroupByOptions
-    from api.types.sequencing_read import SequencingReadGroupByOptions
-    from api.types.sample import SampleGroupByOptions
-    from api.types.host_organism import HostOrganismGroupByOptions
-else:
-    ConsensusGenomeGroupByOptions = strawberry.lazy("api.types.consensus_genome")
-    SequencingReadGroupByOptions = strawberry.lazy("api.types.sequencing_read")
-    SampleGroupByOptions = strawberry.lazy("api.types.sample")
-    HostOrganismGroupByOptions = strawberry.lazy("api.types.host_organism")
+
+
+
 
 
 @strawberry.type
@@ -23,16 +13,36 @@ class HostOrganismGroupByOptions:
     name: Optional[str] = None
     version: Optional[str] = None
 
+@strawberry.type
+class SampleGroupByOptions:
+    collection_id: Optional[int] = None
+    collection_location: Optional[str] = None
+    host_organism: Optional[HostOrganismGroupByOptions] = None
+
+@strawberry.type
+class SequencingReadGroupByOptions:
+    collection_id: Optional[int] = None
+    sample: Optional[SampleGroupByOptions] = None
+
+@strawberry.type
+class ConsensusGenomeGroupByOptions:
+    collection_id: Optional[int] = None
+    # created_at: Optional[datetime.datetime] = None
+    sequence_read: Optional[SequencingReadGroupByOptions] = None
+
 def build_consensus_genome_group_by_output(group_object: Optional[ConsensusGenomeGroupByOptions], keys: list[str], value: Any) -> ConsensusGenomeGroupByOptions:
     """
     Given a list of group by keys, build a nested object to represent the group by clause
     """
     if not group_object:
-        group_object = Annotated[ConsensusGenomeGroupByOptions, strawberry.type(name="ConsensusGenomeGroupByOptions")]
+        group_object = ConsensusGenomeGroupByOptions()
     
     key = keys.pop(0)
     if key == "sequence_read":
-        value = build_sequencing_read_group_by_output(None, keys, value)
+        if getattr(group_object, key):
+            value = build_sequencing_read_group_by_output(getattr(group_object, key), keys, value)
+        else:
+            value = build_sequencing_read_group_by_output(None, keys, value)
     # Add more cases for other nested 1:1 relationships
     
     setattr(group_object, key, value)
@@ -43,11 +53,14 @@ def build_sequencing_read_group_by_output(group_object: Optional[SequencingReadG
     Given a list of group by keys, build a nested object to represent the group by clause
     """
     if not group_object:
-        group_object = Annotated[SequencingReadGroupByOptions, strawberry.type(name="SequencingReadGroupByOptions")]
+        group_object = SequencingReadGroupByOptions()
     
     key = keys.pop(0)
     if key == "sample":
-        value = build_sample_group_by_output(None, keys, value)
+        if getattr(group_object, key):
+            value = build_sample_group_by_output(getattr(group_object, key), keys, value)
+        else:
+            value = build_sample_group_by_output(None, keys, value)
     
     setattr(group_object, key, value)
     return group_object
@@ -57,11 +70,14 @@ def build_sample_group_by_output(group_object: Optional[SampleGroupByOptions], k
     Given a list of group by keys, build a nested object to represent the group by clause
     """
     if not group_object:
-        group_object = Annotated[SampleGroupByOptions, strawberry.type(name="SampleGroupByOptions")]
+        group_object = SampleGroupByOptions()
     
     key = keys.pop(0)
     if key == "host_organism":
-        value = build_host_organism_group_by_output(None, keys, value)
+        if getattr(group_object, key):
+            value = build_host_organism_group_by_output(getattr(group_object, key), keys, value)
+        else:
+            value = build_host_organism_group_by_output(None, keys, value)
     
     setattr(group_object, key, value)
     return group_object
@@ -71,7 +87,6 @@ def build_host_organism_group_by_output(group_object: Optional[HostOrganismGroup
     Given a list of group by keys, build a nested object to represent the group by clause
     """
     if not group_object:
-        # group_object = Annotated[HostOrganismGroupByOptions, strawberry.type(name="HostOrganismGroupByOptions")]
         group_object = HostOrganismGroupByOptions()
     
     key = keys.pop(0)
